// src/payload-types.ts
export interface Media {
  id: string;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface SiteSetting {
  id: string;
  siteName: string;
  defaultMetaTitle: string;
  defaultMetaDescription: string;
  navbarLogo?: (string | null) | Media;
  favicon?: (string | null) | Media;
  navLinks?:
    | {
        label: string;
        href: string;
        openInNewTab?: boolean | null;
        id?: string | null;
      }[]
    | null;
  footerLegalText: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  footerLinks?:
    | {
        label: string;
        href: string;
        id?: string | null;
      }[]
    | null;
  socialLinks?:
    | {
        platform: 'instagram' | 'facebook' | 'youtube' | 'twitter-x';
        url: string;
        id?: string | null;
      }[]
    | null;
  updatedAt?: string | null;
  createdAt?: string | null;
}